pyproject.toml
[tool.poetry]
name = "user-managment-service"
version = "0.1.0"
description = ""
authors = ["Aleksey Cherniaev <cherniaevaleksey2304@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
pytest = "^8.2.2"
bcrypt = "^4.1.3"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
factory-boy = "^3.3.0"
pytest-asyncio = "^0.23.7"
alembic = {extras = ["async"], version = "^1.13.2"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
